requirements.txt
flask
pandas
scikit-learn
nltk
pickle-mixin
python-dotenv
pyzmail36
tldextract